feat: add new step types and code refactoring
