--- app/document/components/ScreenshotView.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-const ScreenshotView = (props: Props) => {
-  return <div>ScreenshotView</div>;
-};
-
-export default ScreenshotView;
