+++ app/document/components/StepType/DirectionStep.tsx
@@ -1,0 +1,344 @@
+import React, { useCallback, useEffect, useRef, useState } from "react";
+import Image from "next/image";
+import { Trash, ImagePlus, Loader, GripVertical } from "lucide-react";
+import CustomAlertDialog from "@/components/ui/Common/CustomAlertDialog";
+import { SyntheticListenerMap } from "@dnd-kit/core/dist/hooks/utilities";
+import { DraggableAttributes } from "@dnd-kit/core";
+import { Screenshot } from "../../document.types";
+
+interface DirectionStepProps {
+  img: string;
+  index: number;
+  info: Screenshot | null;
+  mode: string;
+  stepDescription: string;
+  stepNumber: number;
+  stepId: string;
+  onStepDescriptionChange?: (stepId: string, newDescription: string) => void;
+  handleDeleteStep: (id: string) => void;
+  handleAddNewImage?: (stepNumber: number) => void;
+  handleDeleteImage?: (stepNumber: number) => void;
+  loading?: boolean;
+  dragAttributes?: unknown;
+  dragListeners?: unknown;
+  isDragging?: boolean;
+  annotationColor?: string;
+}
+
+const DirectionStep: React.FC<DirectionStepProps> = ({
+  img,
+  index,
+  info,
+  mode,
+  stepDescription,
+  stepNumber,
+  stepId,
+  onStepDescriptionChange,
+  handleDeleteStep,
+  handleAddNewImage,
+  handleDeleteImage,
+  loading,
+  dragAttributes,
+  dragListeners,
+  isDragging,
+  annotationColor,
+}) => {
+  const [imageDimensions, setImageDimensions] = useState({
+    width: 0,
+    height: 0,
+  });
+  const [displayDimensions, setDisplayDimensions] = useState({
+    width: 0,
+    height: 0,
+  });
+
+  const imgRef = useRef<HTMLImageElement>(null);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const inputRef = useRef<HTMLTextAreaElement>(null);
+  const resizeObserverRef = useRef<ResizeObserver | null>(null);
+  const mutationObserverRef = useRef<MutationObserver | null>(null);
+  const animationFrameRef = useRef<number | null>(null);
+  const originalDescriptionRef = useRef<string>(stepDescription);
+
+  const determineAnnotationColor = (annotationColor: string) => {
+    switch (annotationColor.toLowerCase()) {
+      case "green":
+        return "bg-green-500 border-green-300";
+      case "blue":
+        return "bg-blue-500 border-blue-300";
+      case "yellow":
+        return "bg-yellow-500 border-yellow-300";
+      default:
+        return "bg-blue-500 border-blue-300";
+    }
+  };
+
+  // Handle image load to get original dimensions
+  const handleImageLoad = useCallback(() => {
+    const imgElement = imgRef.current;
+    if (imgElement) {
+      setImageDimensions({
+        width: imgElement.naturalWidth,
+        height: imgElement.naturalHeight,
+      });
+      setDisplayDimensions({
+        width: imgElement.clientWidth,
+        height: imgElement.clientHeight,
+      });
+    }
+  }, []);
+
+  const updateDisplayDimensions = useCallback(() => {
+    if (animationFrameRef.current) {
+      cancelAnimationFrame(animationFrameRef.current);
+    }
+
+    animationFrameRef.current = requestAnimationFrame(() => {
+      const imgElement = imgRef.current;
+      if (imgElement) {
+        setDisplayDimensions({
+          width: imgElement.clientWidth,
+          height: imgElement.clientHeight,
+        });
+      }
+    });
+  }, []);
+
+  const handleDescriptionChange = (
+    e: React.ChangeEvent<HTMLTextAreaElement>
+  ) => {
+    if (mode === "edit") {
+      onStepDescriptionChange?.(stepId, e.target.value);
+    }
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (mode !== "edit") return;
+
+    if (e.key === "Escape") {
+      onStepDescriptionChange?.(stepId, originalDescriptionRef.current);
+      inputRef.current?.blur();
+    } else if (e.key === "Enter") {
+      inputRef.current?.blur();
+    }
+  };
+
+  useEffect(() => {
+    originalDescriptionRef.current = stepDescription;
+  }, [stepDescription]);
+
+  const getResponsivePosition = useCallback(() => {
+    if (imageDimensions.width === 0 || displayDimensions.width === 0) {
+      return { left: "50%", top: "50%" };
+    }
+
+    const captureContext = {
+      devicePixelRatio: info?.devicePixelRatio || 1,
+      viewportWidth: info?.viewportWidth || imageDimensions.width,
+      viewportHeight: info?.viewportHeight || imageDimensions.height,
+    };
+    const coords = {
+      x: info?.viewportX || 0,
+      y: info?.viewportY || 0,
+    };
+
+    if (captureContext) {
+      const { viewportWidth, viewportHeight } = captureContext;
+
+      if (viewportWidth && viewportHeight) {
+        const xPercent = (coords.x / viewportWidth) * 100;
+        const yPercent = (coords.y / viewportHeight) * 100;
+
+        return {
+          left: `${Math.min(Math.max(xPercent, 0), 100)}%`,
+          top: `${Math.min(Math.max(yPercent, 0), 100)}%`,
+        };
+      }
+    }
+
+    const xPercent = (coords.x / imageDimensions.width) * 100;
+    const yPercent = (coords.y / imageDimensions.height) * 100;
+
+    return {
+      left: `${Math.min(Math.max(xPercent, 0), 100)}%`,
+      top: `${Math.min(Math.max(yPercent, 0), 100)}%`,
+    };
+  }, [info, imageDimensions, displayDimensions]);
+
+  const getIndicatorSize = useCallback(() => {
+    if (displayDimensions.width === 0) return 32;
+    const baseSize = 32;
+    const scaleFactor = Math.min(displayDimensions.width / 400, 1.5);
+    return Math.max(16, Math.min(48, baseSize * scaleFactor));
+  }, [displayDimensions.width]);
+
+  useEffect(() => {
+    const imgElement = imgRef.current;
+    if (!imgElement) return;
+
+    if (resizeObserverRef.current) {
+      resizeObserverRef.current.disconnect();
+    }
+
+    resizeObserverRef.current = new ResizeObserver(() => {
+      updateDisplayDimensions();
+    });
+
+    resizeObserverRef.current.observe(imgElement);
+
+    const handleResize = () => updateDisplayDimensions();
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      if (resizeObserverRef.current) {
+        resizeObserverRef.current.disconnect();
+      }
+      window.removeEventListener("resize", handleResize);
+      if (animationFrameRef.current) {
+        cancelAnimationFrame(animationFrameRef.current);
+      }
+    };
+  }, [updateDisplayDimensions]);
+
+  return (
+    <div
+      ref={containerRef}
+      className={`screenshot-item border border-gray-200 rounded-lg p-4 bg-white flex flex-col items-start gap-3 shadow-sm transition-all duration-200 ${
+        isDragging ? "shadow-lg scale-105" : ""
+      }`}
+    >
+      <div className="flex items-start gap-2 text-sm font-medium w-full">
+        {mode === "edit" ? (
+          <span
+            className="cursor-grab active:cursor-grabbing hover:text-blue-800"
+            {...(dragAttributes as DraggableAttributes)}
+            {...(dragListeners as SyntheticListenerMap)}
+          >
+            <GripVertical className="w-4 h-4 inline-block" />{" "}
+            <span className="px-3 py-1 rounded-md font-semibold text-blue-600 bg-blue-100 min-w-24 text-center">
+              step {index + 1}
+            </span>
+          </span>
+        ) : (
+          <span className="px-3 py-1 rounded-md font-semibold text-blue-600 bg-blue-100 min-w-24 text-center">
+            step {index + 1}
+          </span>
+        )}
+        <div className="flex-1">
+          <textarea
+            ref={inputRef}
+            className={`rounded-md w-full h-auto overflow-hidden ${
+              mode === "edit"
+                ? "border-blue-300 bg-white px-2 cursor-text border-none focus:outline-none ring-2 ring-blue-100 focus:ring-2 focus:ring-blue-500"
+                : "font-semibold bg-transparent border-none cursor-default"
+            }`}
+            value={stepDescription}
+            readOnly={mode !== "edit"}
+            disabled={mode !== "edit"}
+            onChange={handleDescriptionChange}
+            onKeyDown={handleKeyDown}
+            placeholder={mode === "edit" ? "Enter step description..." : ""}
+            rows={1}
+            style={{ minHeight: "2.5rem" }}
+            onInput={(e) => {
+              const target = e.target as HTMLTextAreaElement;
+              target.style.height = "auto";
+              target.style.height = `${target.scrollHeight}px`;
+            }}
+          />
+        </div>
+        {mode === "edit" && (
+          <div className="flex items-center gap-2">
+            <div className="p-2 bg-slate-100 rounded-sm cursor-pointer hover:bg-red-100 transition-colors">
+              <CustomAlertDialog
+                title="Delete Step"
+                description={`Are you sure you want to delete step this step? `}
+                onConfirm={() => handleDeleteStep(stepId)}
+                triggerDescription={
+                  <Trash
+                    aria-label="Delete Step"
+                    role="button"
+                    className="w-6 h-6 text-red-500 cursor-pointer hover:text-red-700 transition-colors"
+                  />
+                }
+              />
+            </div>
+          </div>
+        )}
+      </div>
+
+      {img ? (
+        <div className="relative w-full group">
+          <Image
+            ref={imgRef}
+            src={img}
+            alt={`Screenshot ${stepNumber}`}
+            width={info?.viewportWidth || 800}
+            height={info?.viewportHeight || 600}
+            onLoad={handleImageLoad}
+            className={`w-full h-auto border rounded-md transition-all duration-200 ${
+              mode === "edit" ? "group-hover:brightness-80 cursor-pointer" : ""
+            }`}
+            style={{ maxWidth: "100%", height: "auto" }}
+          />
+
+          {mode === "edit" && (
+            <CustomAlertDialog
+              title="Delete Image"
+              description={`Are you sure you want to delete the image from this step? `}
+              onConfirm={() => handleDeleteImage?.(stepNumber)}
+              triggerDescription={
+                <div className="hidden group-hover:flex absolute top-2 right-2 items-center gap-2 bg-white p-1.5 rounded-md shadow-md cursor-pointer hover:shadow-lg transition-all">
+                  <Trash
+                    aria-label="Delete image"
+                    role="button"
+                    className="w-8 h-8 p-2 group-hover:bg-red-400 group-hover:text-slate-50 rounded-full cursor-pointer"
+                  />
+                  <span>Delete image</span>
+                </div>
+              }
+            />
+          )}
+
+          {info?.viewportX !== 0 &&
+            info?.viewportY !== 0 &&
+            imageDimensions.width > 0 &&
+            displayDimensions.width > 0 && (
+              <div
+                className={`absolute opacity-50 border-4 rounded-full bg-opacity-30 transform -translate-x-1/2 -translate-y-1/2 pointer-events-none transition-all duration-200 ${determineAnnotationColor(
+                  annotationColor || "BLUE"
+                )}`}
+                style={{
+                  ...getResponsivePosition(),
+                  width: `${getIndicatorSize()}px`,
+                  height: `${getIndicatorSize()}px`,
+                }}
+                aria-label={`Click indicator for step ${stepNumber}`}
+              >
+                <div className="absolute inset-0 bg-blue-400 rounded-full opacity-50"></div>
+              </div>
+            )}
+        </div>
+      ) : (
+        <>
+          {mode === "edit" && (
+            <div className="w-full h-96 flex items-center justify-center border-2 border-dashed border-gray-300 rounded-lg bg-gray-50">
+              {loading ? (
+                <Loader className="w-12 h-12 text-blue-300 animate-spin" />
+              ) : (
+                <ImagePlus
+                  className="w-24 h-24 text-gray-400 cursor-pointer hover:text-blue-600 transition-colors"
+                  onClick={() => {
+                    handleAddNewImage?.(stepNumber);
+                  }}
+                />
+              )}
+            </div>
+          )}
+        </>
+      )}
+    </div>
+  );
+};
+
+export default DirectionStep;
